+++ src/main/resources/log4j.properties
@@ -10,6 +10,6 @@ log4j.appender.stdout.layout.ConversionPattern=%d{yy/MM/dd HH:mm:ss} %p %c{2}: %
 
 # Set some reasonable defaults
 
-log4j.logger.edu.brown.cs.zookeeper_benchmark=INFO
+log4j.logger.edu.brown.cs.zkbenchmark=INFO
 log4j.logger.com.netflix.curator=ERROR
 log4j.logger.org.apache.zookeeper=ERROR
